Initial commit
This includes all of the basics for a simple terminal-based chess game.
The interface is very boring, but it is sufficient to play around with
the mechanics of the game.

The usage is to type the square of the desired piece and the destination
square separated by a space. For example, moving the queen-side knight
would require entering: `b2 c3`. It is currently case sensitive.

The unit tests are executed using acutest, and can be run via the
makefile using `make test`. The game executable can be run using
`make run`.

Currently it does not support the following:
* en passant
* castling
* pawn promotion
* check (and anything of rules surrounding check)
* ending the game (checkmate, stalemate, draw)