Add support for en passant
There are 3 requirements to be able to perform an en passant attack:
- The attacking pawn must have moved forward 3 ranks
- The victim pawn must have moved forward two spaces on its first move
- The attack must happen on the very next move

To allow for this situation, I have added two variables for squares and
1 for time relating to en passant. When a pawn moves forward two spaces
on its first move, it "enables" en passant by increasing the time
variable, and setting the two square variables to indicate where the
attack must be performed, and which square it will affect. The attack
can only be performed when this time variable is positive, and it will
be decremented at the end of each move (bottoming out at 0). If a pawn
is in the correct position when en passant is "enabled", it can take the
pawn in the victim square on the next move.