Add support for pawn promotion
When a pawn reaches the last rank, it can be promoted to a queen, rook,
bishop, or knight. When the pawn reaches this row, the user is prompted
for which piece to be promoted to. The unit test for this mechanism
requires writing to stdin. It is quite hacky, and likely not portable.