Move to ncurses UI
I have changed from the simple and awful reprinting terminal interface
to a cleaner one uses ncurses. This required several major changes.

In order for the user input to be retrieved inside the `ui` files, the
pawn promotion has been moved out of `moves.c` and into `main.c` where
it can utilize the ui and access the board.
This also meant I had to remove the unit test for pawn promotion since
it now exists inside the ui.

All of the error message that were previously being printed with
`fprintf` to stderr are now being stored in a global `error_message`
buffer which gets temporarily printed in the ui before being cleared
for the next input.

All of the functions that are not being accessed by other c files have
been removed from the headers to make them cleaner.

The UI itself is not perfect. Firstly, the background is black, and it
looks ridiculous with any other color. That means the pieces can't be
black or they won't be visible. I have compromised by making the white
pieces cyan and the black pieces red. The squares could have different
background colors, but because of the blockiness of the pieces, it looks
nicer to only colorize the squares' outlines.

The pieces themselves are based on some I found online and modified. I
like all of them but the knight, but that's a tough shape to fit in
such a small number of characters. The main thing I'm not currently
thrilled with is the user input. All key presses are read in as input,
including backspaces. Until I learn more about different input
techniques, I am stuck with this if I want to print the characters as
they're typed.